Trace image for prefab components are not captured
Trace image for components belonging to a prefab are not uploaded. This was happening because the trace image for prefabs were getting saved to prefabs temp directory instead of the automation process temp directory.
