+++ src/Pixel.Automation.Web.Playwright.Components/ActorComponents/PlaywrightActorComponent.cs
@@ -73,34 +73,7 @@ public abstract class PlaywrightActorComponent : ActorComponent
         }
 
         return (targetControl.ControlName, targetControl.GetApiControl<ILocator>());
-    }
-
-    /// <summary>
-    /// Take a screen shot if capturing screenshot is enabled after Act method finishes
-    /// </summary>
-    /// <returns></returns>
-    public override async Task OnCompletionAsync()
-    {
-        if (TraceManager.IsEnabled)
-        {
-            await CaptureScreenShotAsync();
-        }
-    }
-
-    /// <summary>
-    /// Capture screenshot of the active page
-    /// </summary>
-    /// <returns></returns>
-    public async Task CaptureScreenShotAsync()
-    {
-        var ownerApplicationEntity = this.EntityManager.GetApplicationEntity(this);
-        if (TraceManager.IsEnabled && ownerApplicationEntity.AllowCaptureScreenshot)
-        {
-            string imageFile = Path.Combine(this.EntityManager.GetCurrentFileSystem().TempDirectory, $"{Path.GetRandomFileName()}.jpeg");
-            await ownerApplicationEntity.CaptureScreenShotAsync(imageFile);
-            TraceManager.AddImage(Path.GetFileName(imageFile));
-        }     
-    }
+    } 
 
     protected void ThrowIfMissingControlEntity()
     {
