+++ src/Pixel.Automation.Web.Selenium.Components/ActorComponents/SeleniumActorComponent.cs
@@ -37,34 +37,7 @@ public abstract class SeleniumActorComponent : ActorComponent
     protected SeleniumActorComponent(string name = "", string tag = "") : base(name, tag)
     {
 
-    }
-
-    /// <summary>
-    /// Take a screen shot if capturing screenshot is enabled after Act method finishes
-    /// </summary>
-    /// <returns></returns>
-    public override async Task OnCompletionAsync()
-    {
-        if (TraceManager.IsEnabled)
-        {
-            await CaptureScreenShotAsync();
-        }
-    }
-
-    /// <summary>
-    /// Capture screenshot of the active page
-    /// </summary>
-    /// <returns></returns>
-    public async Task CaptureScreenShotAsync()
-    {
-        var ownerApplicationEntity = this.EntityManager.GetApplicationEntity(this);
-        if(TraceManager.IsEnabled && ownerApplicationEntity.AllowCaptureScreenshot)
-        {
-            string imageFile = Path.Combine(this.EntityManager.GetCurrentFileSystem().TempDirectory, $"{Path.GetRandomFileName()}.jpeg");
-            await ownerApplicationEntity.CaptureScreenShotAsync(imageFile);
-            TraceManager.AddImage(Path.GetFileName(imageFile));
-        }       
-    }
+    }    
 }
 
 /// <summary>
